requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.3.1+cpu
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
distro==1.9.0
exceptiongroup==1.3.0
faiss-cpu==1.11.0
fastapi==0.115.12
filelock==3.13.1
fsspec==2024.6.1
h11==0.16.0
hf_xet==1.1.2
httpcore==1.0.9
httpx==0.25.2
huggingface-hub==0.31.2
idna==3.10
Jinja2==3.1.4
jiter==0.9.0
joblib==1.5.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.2.6
openai==1.79.0
packaging==25.0
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
pytest==8.3.5
python-dotenv==1.1.0
python-telegram-bot==20.6
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
